Major updates across the board.
I should've been committing more often, oops.

The big thing that I changed was that filter_mass was not being properly
compiled with MPI.  This caused nonsense results when trying to run in
parallel.

Currently there is an inconsistency for subhalos.  Within SAGE whenever
a galaxy is eovlved, the only halo it uses is the main FOF halo.
However, there can (and is) cases where the FOF halo does not have its
infalled suppressed but the satellite halo does!  As a result the
subhalo gets added to the list but since it never gets "processed" in
SAGE (only the main FOF halo does), its baryonic accretion is not
suppressed.  Fix or not?