Updated core_save.c to use macros.
Saving of grids is now much cleaner.  There are two macros defined, one
for saving the property directly and one if a conversion needs to be
made.  This second macro is necessary because some properties such as
SFR have weird internal units but need to be saved as Msun/yr.