Fixed some parts of the dust model.
Previously I was not properly initializing the ejected dust reservoir
causing NaNs to pop up.  Furthermore, my units were out of whack;
specifically the value of Ms and yd need to be converted to SAGE units
of 1e10 Msun/h.

The results are a bit perplexing.  The total dust mass (hot + cold +
ejected) as a function of stellar mass seems to be correct but there is
a ridiculously large amount of scatter for the total dust mass as a
function of halo virial mass.  This is because for a given virial mass
there is a wide range of galaxy stellar masses.  Unsure if this is
physical or a bug...