Started reworking SAGE malloc scheme.
Have gotten rid of the old table of pointers and replaced with just
normal allocation.  As a result, if I want to keep track of the memory I
use, I need to specify the size of the memory that is being freed by
each myfree() call.

I need to go through the code and replace all calls to malloc() with
calls to mymalloc.  Furthermore, I may need to instead use calloc()
within mymalloc.  If this is the case, I would need to call mymalloc
with the length and size of each element (i.e., mymalloc(MAXSNAPS,
sizeof(int)) rather than mymalloc(MAXSNAPS * sizeof(int))).