Dust model is now working properly.
The issue was that I was plotting ALL galaxies, including both centrals
and satellites.  When I plotted the "Virial Mass" this was the virial
mass of the FoF Halo, not the virial mass of the host halo.  This would
cause a discrepancy for satellite galaxies whose stellar masses are
small compared to the virial mass of the entire FoF Halo (but still
correct if you compared them to the virial mass of the subhalo).