Updated memory manager in SAGE.
Adjust the calls to mymalloc to now be calls to mycalloc.  The signature
for this function is identical to calloc (i.e., mycalloc(nmemb, size)).
Since we don't track the memory in a table of pointers, we don't know
the size of memory as we free it.  So tracking of memory can be done
properly, the function signature of `myfree` has been adjusted; `myfree`
is now called as `myfree(data_to_be_freed, data_size)`.

Valgrind checked with no leaks.