Started folding in the gridding into SAGE.
For the self_consistent runs, since we are only creating a single grid
of ionizing photons, we can do this natively within SAGE.  This will
decrease runtime as I don't have to reopen all the files again and place
them inside the grids. This will add an extra memory requirement to SAGE
(0.13GB) but as it seems that SAGE itself only takes ~50 MB to run, this
hopefully will still be feasible.