Fixed the slowdown problem.
Turns out the issue was with the calloc for the innermost loop of
load_trees.  Since calloc initializes all allocated memory to 0, the
computation time was being bloated by it.  Since the loop was being
called for each tree, this lead to immense slowdown.

I have since changed from calloc to malloc in the innermost loop and the
code is now running smoothly as before.