Work on optimizations.
Have moved the calculation of coreburning and Delta_Eta to use look-up
tables.  I calculate this for a range of values at the start of the run
and then find the correct index when it needs to be computed.

Run time has dropped from ~60s for File 0 to ~40s for File 0.