Touched up the lookup-table calculation. Also added test.
There was an index overflow when calling calculate_Delta_Eta with m_high
= 120.0.  This has been fixed by checking for this case and subtracting
one from the bin index.

I've added two galaxy files to the test_RSAGE repository
(https://github.com/jacobseiler/test_RSAGE) for comparisons testing.
I've added a test to test_SAGE.py that compares the masses of the test
galaxies and these data galaxies.  An error is raised if the difference
in masses is greater than 5e-3.