Started adding tracking of stellar ages.
This tracking will be used to determine the number of ionizing photons
emitted using the stellar mass rather than the instantaneous SFR.

Since I am already tracking previous SF episodes for the delayed SN,
this commit goes through and renames variables for the delayed SN.  This
is because we could want to do IRA but do the tracking of stellar ages
(and vice versa).

Furthermore, we may also wish to have different resolutions.  Only need
past ~50Myrs of SF for delayed SN but we may need more/less for stellar
tracking...