Fixed the error.
The problem was with the delayedSN feedback.  I was passing False
instead of True for ``update_stars`` meaning the delayedSN wasn't
actually doing anything!

I have fixed this and other bugs with the Stellar Tracking.  The stellar
tracking now compiles and runs but I have yet to check the results and
confirm that they're similar order of magnitude to the SFR case.