Attempt to reduce computation time of Cooling calculations.
The runtime of SAGE is heavily dominated by calculating cooling rates.
In an attempt to reduce this, I created a lookup table that uses the
temperature and metallicity of the galaxy.

Such a modification does not change the SMF but only reduces the
computation time for file 0 by 3 seconds (previous calls of
cooling/heating took 9 seconds and now 6 seconds).

I need to properly check vtune to see where the hotspots are for the new
function.