IGNORE LAST COMMIT.
I forgot to update the run of test_sage to actually point to the new
SAGE executable.  Testing properly yields a different number of galaxies
with nonzero masses and the galaxy masses can differ by ~0.5dex.

I believe this is because of the numerical precision of the calculation.
In my lookup table I'm calculating in log bins of 0.01.  However in
SAGE, galaxies differ by much more than this.  Hence since this function
can be quite non-linear, binning into these "coarse" bins yields vastly
different cooling rates.

I am going to have to shelve this idea of using a metal lookup table for
now.  It seems that I'll have to find another way to optimize the
cooling functions.