Update on the reionization_modifier for satellites (#6).
This currently does not pass `test_sage.py`.  Because we are no longer recomputing the
`reionization_modifier` for satellites, the number of baryons in the satellite (and hence
that are stripped) is different.  The mean difference in stellar mass (compared to the test_RSAGE
repo) is ~0.02dex for galaxies whose masses move by more than 0.003 and 0.001 for all galaxies.