Redid the reduction of the gas through the ejection channels.
Previously, I subtracted an equal amounts from both the SN and QSO channels.  However
this doesn't really make physical sense.  Consider if the ejected reservoir was composed
of 90% SN gas and 10% QSO gas.  During a reincorporation/infall event it then doesn't make
sense to give an equal weight to both components.

Instead what I've doen is now decrease the channels in proportion to the ratio of SN/QSO gas
in the ejected reservoirs.  This commit readded the XASSERT statements and the tests pass.